배열의 특징

+++ src/generic/ex4/MethodMain1.java
@@ -5,7 +5,7 @@ import generic.animal.Dog;
 public class MethodMain1 {
     
     public static void main(String[] args) {
-        Integer i = 10;
+        Integer i = (Integer) 10;
         
         Integer o = (Integer) GenericMethod.objMethod(i); // 캐스팅 해야함
         
@@ -13,7 +13,7 @@ public class MethodMain1 {
         System.out.println("명시적 타입 인자 전달");
         Integer result = GenericMethod.<Integer>genericMethod(i);// 타입인자 전달
         Integer integerValue = GenericMethod.<Integer>numberMethod(i);
-        Double aDouble = GenericMethod.<Double>numberMethod(20.0);
+        Double aDouble = GenericMethod.<Double>numberMethod(Double.valueOf(20.0));
         
         GenericMethod genericMethod = new GenericMethod();
         Dog gd = genericMethod.test(new Dog("gd", 100));
